scripts/requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0
rich>=13.0
cryptography>=41.0
watchdog>=4.0
mcp>=1.0
pyyaml>=6.0
